+++ src/process/validation/util.ts
@@ -1,7 +1,22 @@
 import { FieldError } from 'error';
+import { Component } from 'types';
 import { Evaluator, unescapeHTML } from 'utils';
 import { VALIDATION_ERRORS } from './i18n';
-import { Component } from 'types';
+import _isEmpty from 'lodash/isEmpty';
+import _isObject from 'lodash/isObject';
+import _isPlainObject from 'lodash/isPlainObject';
+
+export function isComponentPersistent(component: Component) {
+  return component.persistent ? component.persistent : true;
+}
+
+export function isComponentProtected(component: Component) {
+  return component.protected ? component.protected : false;
+}
+
+export function isEmptyObject(obj: any) {
+  return !!obj && Object.keys(obj).length === 0 && obj.constructor === Object;
+}
 
 export function toBoolean(value: any) {
   switch (typeof value) {
@@ -29,8 +44,8 @@ export function isPromise(value: any): value is Promise<any> {
   );
 }
 
-export function keepErrorSettingInSecret(rule: string, component: Component): boolean {
-  return rule === 'custom' && !!component.validate?.customPrivate;
+export function isObject(obj: any): obj is object {
+  return obj != null && (typeof obj === 'object' || typeof obj === 'function');
 }
 
 const getCustomErrorMessage = ({ errorKeyOrMessage, context }: FieldError): string =>
@@ -69,9 +84,35 @@ export const interpolateErrors = (errors: FieldError[], lang: string = 'en') => 
         label: context.component.label || context.component.placeholder || context.component.key,
         path: context.path,
         value: context.value,
-        setting: keepErrorSettingInSecret(error.ruleName, context.component) ? '' : context.setting,
+        setting: context.setting,
         index: context.index || 0,
       },
     };
   });
+};
+
+export const hasValue = (value: any) => {
+  if (_isObject(value)) {
+    return !_isEmpty(value);
+  }
+
+  return (typeof value === 'number' && !Number.isNaN(value)) || !!value;
+};
+
+export const doesArrayDataHaveValue = (dataValue: any[] = []): boolean => {
+  if (!Array.isArray(dataValue)) {
+    return !!dataValue;
+  }
+
+  if (!dataValue.length) {
+    return false;
+  }
+
+  const isArrayDataComponent = dataValue.every(_isPlainObject);
+
+  if (isArrayDataComponent) {
+    return dataValue.some((value) => Object.values(value).some(hasValue));
+  }
+
+  return dataValue.some(hasValue);
 };