+++ src/types/process/ProcessContext.ts
@@ -6,7 +6,6 @@ import {
   ProcessType,
   ProcessorInfo,
   ComponentPaths,
-  LocalRoot,
 } from 'types';
 
 export type ComponentInstances = {
@@ -24,7 +23,6 @@ export type BaseProcessContext<ProcessorScope> = {
   submission?: any;
   flat?: boolean;
   parent?: Component;
-  localRoot?: LocalRoot;
   parentPaths?: ComponentPaths;
   local?: boolean; // If the "data" being passed to the processors is local to the nested form.
   evalContext?: (context: ProcessorContext<ProcessorScope>) => any;